chore: theme only exports mainTheme, removed useless barrel
